requirements.txt
aiogram==3.27.0
fastapi==0.135.3
uvicorn==0.44.0
openai==2.32.0
aiohttp==3.13.5
python-dotenv==1.1.0
